+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 ### Removed `stdout` and `stderr` logs for Cromwell, since their contents are included in the `backend log`.  Also increased the maximum amount of content for the `backend log` to 500KB.
 
+### No longer treat failures due to an attempt being preempted in Cromwell as failures.
+
 ### All attempts (not just the latest) are now included in the timing diagram tab on the Job Details page.
 
 ## v1.5.1 Release Notes